+++ src/watchtracker/__init__.py
@@ -1,3 +1,3 @@
 """Personal Media Tracker."""
 
-__version__ = "2.1.2"
+__version__ = "2.1.3"